requirements.txt
Django==2.2
djangorestframework==3.13.1
mysqlclient==1.4.6
PyMySQL==0.9.3
PyYAML==5.3.1
requests==2.24.0
uWSGI==2.0.21
